Development v2

+++ src/configs/axios.js
@@ -26,13 +26,18 @@ instance.interceptors.response.use(
   },
   async error => {
     let originalConfig = error.config
-    const { status, data } = error.response
-    if (originalConfig.url !== '/user/login') {
-      if (status === 401 && data.message === 'jwt expired') {
-        window.localStorage.clear()
-        window.location.href = '/login'
+    
+    // Check if error.response exists before destructuring
+    if (error.response) {
+      const { status, data } = error.response
+      if (originalConfig.url !== '/user/login') {
+        if (status === 401 && data.message === 'jwt expired') {
+          window.localStorage.clear()
+          window.location.href = '/login'
+        }
       }
     }
+    
     return Promise.reject(error)
   }
 )
